Generate hook-directory contents
As it turns out, hooks in pbuilder don't know a lot about their configuration.
In particular, they don't know about the value of $BUILDRESULT, which is impor-
tant, if we'd like to make the D05resultsaredeps portable.

This fixes the situation by interleaving the value of that configuration para-
meter into the generated script which in turn is used by pbuilder.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>